--- lib/brokers/common.js
@@ -1,20 +1,0 @@
-'use strict'
-function newChannels () {
-  const state = {}
-  return {
-    set: (c) => {
-      state[c.id] = c
-    },
-    getById: (id) => {
-      return state[id]
-    },
-    removeById: (id) => {
-      delete state[id]
-    },
-    forEach: (fun) => {
-      Object.values(state).forEach(fun)
-    }
-  }
-}
-
-module.exports = { newChannels }
