--- lib/brokers/index.js
@@ -1,7 +1,0 @@
-'use strict'
-
-const rpcBroker = require('./rpc')
-
-module.exports = {
-  rpc: rpcBroker.create
-}
