+++ lib/brokers/rpc.js
@@ -4,7 +4,24 @@
 const yac = require('ya-common')
 const logger = yac.logger(__filename)
 const { COMMANDS } = yac.constants
-const { newChannels } = require('./common')
+
+function newChannels () {
+  const state = {}
+  return {
+    set: (c) => {
+      state[c.id] = c
+    },
+    getById: (id) => {
+      return state[id]
+    },
+    removeById: (id) => {
+      delete state[id]
+    },
+    forEach: (fun) => {
+      Object.values(state).forEach(fun)
+    }
+  }
+}
 
 function newExecutionContext () {
   const state = []
@@ -86,6 +103,18 @@ function newRequests () {
   }
 }
 
+/**
+ * A number, or a string containing a number.
+ * @typedef {(number|string)} NumberLike
+ */
+
+
+/**
+ * Add two numbers together
+ * @param  {Number} num1 The first number
+ * @param  {Number} num2 The second number
+ * @return {NumberLike}      The total of the two numbers
+ */
 function create () {
   const channels = newChannels()
   const execContext = newExecutionContext()
@@ -116,7 +145,6 @@ function create () {
     // logger.debug('executeProcedures')
     while (executeProcedure()) { }
   }
-
   function schedule (channel, id, procedure, args, affinity = null, withStatus = false, withProgress = false) {
     // logger.debug(`execute requestor:${channel.id} ${procedure} ${affinity} ${withStatus}`)
     requests.add(id, channel.id, procedure, args, affinity, withStatus, withProgress)
